refactor(ui): extract shared ColorGridPicker for the accent picker
Per review on #7009: move the accent swatch grid out of GeneralSection into a
reusable core/ui/ColorGridPicker (swatch trigger + popover + radiogroup grid
with an optional default/unset icon cell + a footer slot). GeneralSection now
renders <ColorGridPicker> and passes the presets, default option and the
custom-colour ColorInput as the footer.

Removes the three raw-<button> eslint-disables from GeneralSection; the raw
elements now live in the DS layer (core/ui/), which the lint rule exempts.
Markup/styles preserved verbatim, so the control is unchanged visually.
